+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.5.1] - 2018-10-27
+
+### Added
+
+- An opt-in "exit" Cargo feature to have the panic handler perform an exit
+  semihosting call after logging the panic message.
+
 ## [v0.5.0] - 2018-09-10
 
 - [breaking-change] The `panic_handler` feature gate has been removed. This
@@ -40,7 +47,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial release
 
-[Unreleased]: https://github.com/rust-embedded/panic-semihosting/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/rust-embedded/panic-semihosting/compare/v0.5.1...HEAD
+[v0.5.1]: https://github.com/rust-embedded/panic-semihosting/compare/v0.5.0...v0.5.1
 [v0.5.0]: https://github.com/rust-embedded/panic-semihosting/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/rust-embedded/panic-semihosting/compare/v0.3.0...v0.4.0
 [v0.3.0]: https://github.com/rust-embedded/panic-semihosting/compare/v0.2.0...v0.3.0